// src/app/models/property.model.ts
export interface Properties {
    Id: number;
    UserId: number;
    Name: string;
    Price: string;
    Description: string;
    ContactNo: string;
    PropertySize: string;
    PropertyConfiguration: string;
    State: string;
    District: string;
    Address: string;
    Aminities: string;
    LocationURL: string;
  }
  